@@ -14,11 +14,16 @@ android {
         exclude 'META-INF/NOTICE.txt'
     }
 
-    compileSdkVersion 30
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    compileSdkVersion 31
     defaultConfig {
         applicationId "com.restart.spacestationtracker"
         minSdkVersion 17
-        targetSdkVersion 30
+        targetSdkVersion 31
         versionCode 31
         versionName "5.01"
         testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
@@ -35,23 +40,26 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    testImplementation 'junit:junit:4.13.1'
+    testImplementation 'junit:junit:4.13.2'
 
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
-    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'androidx.appcompat:appcompat:1.3.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.1'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
     implementation 'androidx.cardview:cardview:1.0.0'
-    implementation 'com.google.android.material:material:1.2.1'
+    implementation 'com.google.android.material:material:1.4.0'
     implementation "androidx.coordinatorlayout:coordinatorlayout:1.1.0"
 
-    implementation 'com.google.firebase:firebase-core:18.0.0'
-    implementation 'com.google.firebase:firebase-ads:19.6.0'
+    implementation platform('com.google.firebase:firebase-bom:28.4.2')
+
+    implementation 'com.google.firebase:firebase-core:19.0.2'
+    implementation 'com.google.firebase:firebase-ads:20.4.0'
     implementation 'com.google.firebase:firebase-crash:16.2.1'
-    implementation 'com.google.android.gms:play-services-maps:17.0.0'
-    implementation 'com.google.android.gms:play-services-location:17.1.0'
+    implementation 'com.google.android.gms:play-services-maps:17.0.1'
+    implementation 'com.google.android.gms:play-services-location:18.0.0'
+    implementation 'com.google.firebase:firebase-analytics'
 
     implementation 'com.nightonke:boommenu:2.1.1'
-    implementation 'com.android.volley:volley:1.1.0'
+    implementation 'com.android.volley:volley:1.2.1'
     implementation 'com.github.wooplr:Spotlight:1.2.3'
     implementation 'de.hdodenhof:circleimageview:2.2.0'
     implementation 'com.github.ksoichiro:android-observablescrollview:1.6.0'
